+++ dwds/debug_extension/web/data_types.g.dart
@@ -7,12 +7,12 @@ part of 'data_types.dart';
 // **************************************************************************
 
 Serializer<ConnectFailure> _$connectFailureSerializer =
-    new _$ConnectFailureSerializer();
+    _$ConnectFailureSerializer();
 Serializer<DevToolsOpener> _$devToolsOpenerSerializer =
-    new _$DevToolsOpenerSerializer();
-Serializer<DevToolsUrl> _$devToolsUrlSerializer = new _$DevToolsUrlSerializer();
+    _$DevToolsOpenerSerializer();
+Serializer<DevToolsUrl> _$devToolsUrlSerializer = _$DevToolsUrlSerializer();
 Serializer<DebugStateChange> _$debugStateChangeSerializer =
-    new _$DebugStateChangeSerializer();
+    _$DebugStateChangeSerializer();
 
 class _$ConnectFailureSerializer
     implements StructuredSerializer<ConnectFailure> {
@@ -22,11 +22,8 @@ class _$ConnectFailureSerializer
   final String wireName = 'ConnectFailure';
 
   @override
-  Iterable<Object?> serialize(
-    Serializers serializers,
-    ConnectFailure object, {
-    FullType specifiedType = FullType.unspecified,
-  }) {
+  Iterable<Object?> serialize(Serializers serializers, ConnectFailure object,
+      {FullType specifiedType = FullType.unspecified}) {
     final result = <Object?>[
       'tabId',
       serializers.serialize(object.tabId, specifiedType: const FullType(int)),
@@ -36,20 +33,17 @@ class _$ConnectFailureSerializer
     if (value != null) {
       result
         ..add('reason')
-        ..add(
-          serializers.serialize(value, specifiedType: const FullType(String)),
-        );
+        ..add(serializers.serialize(value,
+            specifiedType: const FullType(String)));
     }
     return result;
   }
 
   @override
   ConnectFailure deserialize(
-    Serializers serializers,
-    Iterable<Object?> serialized, {
-    FullType specifiedType = FullType.unspecified,
-  }) {
-    final result = new ConnectFailureBuilder();
+      Serializers serializers, Iterable<Object?> serialized,
+      {FullType specifiedType = FullType.unspecified}) {
+    final result = ConnectFailureBuilder();
 
     final iterator = serialized.iterator;
     while (iterator.moveNext()) {
@@ -58,20 +52,12 @@ class _$ConnectFailureSerializer
       final Object? value = iterator.current;
       switch (key) {
         case 'tabId':
-          result.tabId =
-              serializers.deserialize(
-                    value,
-                    specifiedType: const FullType(int),
-                  )!
-                  as int;
+          result.tabId = serializers.deserialize(value,
+              specifiedType: const FullType(int))! as int;
           break;
         case 'reason':
-          result.reason =
-              serializers.deserialize(
-                    value,
-                    specifiedType: const FullType(String),
-                  )
-                  as String?;
+          result.reason = serializers.deserialize(value,
+              specifiedType: const FullType(String)) as String?;
           break;
       }
     }
@@ -88,17 +74,12 @@ class _$DevToolsOpenerSerializer
   final String wireName = 'DevToolsOpener';
 
   @override
-  Iterable<Object?> serialize(
-    Serializers serializers,
-    DevToolsOpener object, {
-    FullType specifiedType = FullType.unspecified,
-  }) {
+  Iterable<Object?> serialize(Serializers serializers, DevToolsOpener object,
+      {FullType specifiedType = FullType.unspecified}) {
     final result = <Object?>[
       'newWindow',
-      serializers.serialize(
-        object.newWindow,
-        specifiedType: const FullType(bool),
-      ),
+      serializers.serialize(object.newWindow,
+          specifiedType: const FullType(bool)),
     ];
 
     return result;
@@ -106,11 +87,9 @@ class _$DevToolsOpenerSerializer
 
   @override
   DevToolsOpener deserialize(
-    Serializers serializers,
-    Iterable<Object?> serialized, {
-    FullType specifiedType = FullType.unspecified,
-  }) {
-    final result = new DevToolsOpenerBuilder();
+      Serializers serializers, Iterable<Object?> serialized,
+      {FullType specifiedType = FullType.unspecified}) {
+    final result = DevToolsOpenerBuilder();
 
     final iterator = serialized.iterator;
     while (iterator.moveNext()) {
@@ -119,12 +98,8 @@ class _$DevToolsOpenerSerializer
       final Object? value = iterator.current;
       switch (key) {
         case 'newWindow':
-          result.newWindow =
-              serializers.deserialize(
-                    value,
-                    specifiedType: const FullType(bool),
-                  )!
-                  as bool;
+          result.newWindow = serializers.deserialize(value,
+              specifiedType: const FullType(bool))! as bool;
           break;
       }
     }
@@ -140,11 +115,8 @@ class _$DevToolsUrlSerializer implements StructuredSerializer<DevToolsUrl> {
   final String wireName = 'DevToolsUrl';
 
   @override
-  Iterable<Object?> serialize(
-    Serializers serializers,
-    DevToolsUrl object, {
-    FullType specifiedType = FullType.unspecified,
-  }) {
+  Iterable<Object?> serialize(Serializers serializers, DevToolsUrl object,
+      {FullType specifiedType = FullType.unspecified}) {
     final result = <Object?>[
       'tabId',
       serializers.serialize(object.tabId, specifiedType: const FullType(int)),
@@ -156,12 +128,9 @@ class _$DevToolsUrlSerializer implements StructuredSerializer<DevToolsUrl> {
   }
 
   @override
-  DevToolsUrl deserialize(
-    Serializers serializers,
-    Iterable<Object?> serialized, {
-    FullType specifiedType = FullType.unspecified,
-  }) {
-    final result = new DevToolsUrlBuilder();
+  DevToolsUrl deserialize(Serializers serializers, Iterable<Object?> serialized,
+      {FullType specifiedType = FullType.unspecified}) {
+    final result = DevToolsUrlBuilder();
 
     final iterator = serialized.iterator;
     while (iterator.moveNext()) {
@@ -170,20 +139,12 @@ class _$DevToolsUrlSerializer implements StructuredSerializer<DevToolsUrl> {
       final Object? value = iterator.current;
       switch (key) {
         case 'tabId':
-          result.tabId =
-              serializers.deserialize(
-                    value,
-                    specifiedType: const FullType(int),
-                  )!
-                  as int;
+          result.tabId = serializers.deserialize(value,
+              specifiedType: const FullType(int))! as int;
           break;
         case 'url':
-          result.url =
-              serializers.deserialize(
-                    value,
-                    specifiedType: const FullType(String),
-                  )!
-                  as String;
+          result.url = serializers.deserialize(value,
+              specifiedType: const FullType(String))! as String;
           break;
       }
     }
@@ -200,39 +161,31 @@ class _$DebugStateChangeSerializer
   final String wireName = 'DebugStateChange';
 
   @override
-  Iterable<Object?> serialize(
-    Serializers serializers,
-    DebugStateChange object, {
-    FullType specifiedType = FullType.unspecified,
-  }) {
+  Iterable<Object?> serialize(Serializers serializers, DebugStateChange object,
+      {FullType specifiedType = FullType.unspecified}) {
     final result = <Object?>[
       'tabId',
       serializers.serialize(object.tabId, specifiedType: const FullType(int)),
       'newState',
-      serializers.serialize(
-        object.newState,
-        specifiedType: const FullType(String),
-      ),
+      serializers.serialize(object.newState,
+          specifiedType: const FullType(String)),
     ];
     Object? value;
     value = object.reason;
     if (value != null) {
       result
         ..add('reason')
-        ..add(
-          serializers.serialize(value, specifiedType: const FullType(String)),
-        );
+        ..add(serializers.serialize(value,
+            specifiedType: const FullType(String)));
     }
     return result;
   }
 
   @override
   DebugStateChange deserialize(
-    Serializers serializers,
-    Iterable<Object?> serialized, {
-    FullType specifiedType = FullType.unspecified,
-  }) {
-    final result = new DebugStateChangeBuilder();
+      Serializers serializers, Iterable<Object?> serialized,
+      {FullType specifiedType = FullType.unspecified}) {
+    final result = DebugStateChangeBuilder();
 
     final iterator = serialized.iterator;
     while (iterator.moveNext()) {
@@ -241,28 +194,16 @@ class _$DebugStateChangeSerializer
       final Object? value = iterator.current;
       switch (key) {
         case 'tabId':
-          result.tabId =
-              serializers.deserialize(
-                    value,
-                    specifiedType: const FullType(int),
-                  )!
-                  as int;
+          result.tabId = serializers.deserialize(value,
+              specifiedType: const FullType(int))! as int;
           break;
         case 'newState':
-          result.newState =
-              serializers.deserialize(
-                    value,
-                    specifiedType: const FullType(String),
-                  )!
-                  as String;
+          result.newState = serializers.deserialize(value,
+              specifiedType: const FullType(String))! as String;
           break;
         case 'reason':
-          result.reason =
-              serializers.deserialize(
-                    value,
-                    specifiedType: const FullType(String),
-                  )
-                  as String?;
+          result.reason = serializers.deserialize(value,
+              specifiedType: const FullType(String)) as String?;
           break;
       }
     }
@@ -278,19 +219,15 @@ class _$ConnectFailure extends ConnectFailure {
   final String? reason;
 
   factory _$ConnectFailure([void Function(ConnectFailureBuilder)? updates]) =>
-      (new ConnectFailureBuilder()..update(updates))._build();
+      (ConnectFailureBuilder()..update(updates))._build();
 
-  _$ConnectFailure._({required this.tabId, this.reason}) : super._() {
-    BuiltValueNullFieldError.checkNotNull(tabId, r'ConnectFailure', 'tabId');
-  }
-
+  _$ConnectFailure._({required this.tabId, this.reason}) : super._();
   @override
   ConnectFailure rebuild(void Function(ConnectFailureBuilder) updates) =>
       (toBuilder()..update(updates)).build();
 
   @override
-  ConnectFailureBuilder toBuilder() =>
-      new ConnectFailureBuilder()..replace(this);
+  ConnectFailureBuilder toBuilder() => ConnectFailureBuilder()..replace(this);
 
   @override
   bool operator ==(Object other) {
@@ -344,7 +281,6 @@ class ConnectFailureBuilder
 
   @override
   void replace(ConnectFailure other) {
-    ArgumentError.checkNotNull(other, 'other');
     _$v = other as _$ConnectFailure;
   }
 
@@ -357,14 +293,10 @@ class ConnectFailureBuilder
   ConnectFailure build() => _build();
 
   _$ConnectFailure _build() {
-    final _$result =
-        _$v ??
-        new _$ConnectFailure._(
+    final _$result = _$v ??
+        _$ConnectFailure._(
           tabId: BuiltValueNullFieldError.checkNotNull(
-            tabId,
-            r'ConnectFailure',
-            'tabId',
-          ),
+              tabId, r'ConnectFailure', 'tabId'),
           reason: reason,
         );
     replace(_$result);
@@ -377,23 +309,15 @@ class _$DevToolsOpener extends DevToolsOpener {
   final bool newWindow;
 
   factory _$DevToolsOpener([void Function(DevToolsOpenerBuilder)? updates]) =>
-      (new DevToolsOpenerBuilder()..update(updates))._build();
+      (DevToolsOpenerBuilder()..update(updates))._build();
 
-  _$DevToolsOpener._({required this.newWindow}) : super._() {
-    BuiltValueNullFieldError.checkNotNull(
-      newWindow,
-      r'DevToolsOpener',
-      'newWindow',
-    );
-  }
-
+  _$DevToolsOpener._({required this.newWindow}) : super._();
   @override
   DevToolsOpener rebuild(void Function(DevToolsOpenerBuilder) updates) =>
       (toBuilder()..update(updates)).build();
 
   @override
-  DevToolsOpenerBuilder toBuilder() =>
-      new DevToolsOpenerBuilder()..replace(this);
+  DevToolsOpenerBuilder toBuilder() => DevToolsOpenerBuilder()..replace(this);
 
   @override
   bool operator ==(Object other) {
@@ -412,7 +336,8 @@ class _$DevToolsOpener extends DevToolsOpener {
   @override
   String toString() {
     return (newBuiltValueToStringHelper(r'DevToolsOpener')
-      ..add('newWindow', newWindow)).toString();
+          ..add('newWindow', newWindow))
+        .toString();
   }
 }
 
@@ -437,7 +362,6 @@ class DevToolsOpenerBuilder
 
   @override
   void replace(DevToolsOpener other) {
-    ArgumentError.checkNotNull(other, 'other');
     _$v = other as _$DevToolsOpener;
   }
 
@@ -450,14 +374,10 @@ class DevToolsOpenerBuilder
   DevToolsOpener build() => _build();
 
   _$DevToolsOpener _build() {
-    final _$result =
-        _$v ??
-        new _$DevToolsOpener._(
+    final _$result = _$v ??
+        _$DevToolsOpener._(
           newWindow: BuiltValueNullFieldError.checkNotNull(
-            newWindow,
-            r'DevToolsOpener',
-            'newWindow',
-          ),
+              newWindow, r'DevToolsOpener', 'newWindow'),
         );
     replace(_$result);
     return _$result;
@@ -471,19 +391,15 @@ class _$DevToolsUrl extends DevToolsUrl {
   final String url;
 
   factory _$DevToolsUrl([void Function(DevToolsUrlBuilder)? updates]) =>
-      (new DevToolsUrlBuilder()..update(updates))._build();
+      (DevToolsUrlBuilder()..update(updates))._build();
 
-  _$DevToolsUrl._({required this.tabId, required this.url}) : super._() {
-    BuiltValueNullFieldError.checkNotNull(tabId, r'DevToolsUrl', 'tabId');
-    BuiltValueNullFieldError.checkNotNull(url, r'DevToolsUrl', 'url');
-  }
-
+  _$DevToolsUrl._({required this.tabId, required this.url}) : super._();
   @override
   DevToolsUrl rebuild(void Function(DevToolsUrlBuilder) updates) =>
       (toBuilder()..update(updates)).build();
 
   @override
-  DevToolsUrlBuilder toBuilder() => new DevToolsUrlBuilder()..replace(this);
+  DevToolsUrlBuilder toBuilder() => DevToolsUrlBuilder()..replace(this);
 
   @override
   bool operator ==(Object other) {
@@ -534,7 +450,6 @@ class DevToolsUrlBuilder implements Builder<DevToolsUrl, DevToolsUrlBuilder> {
 
   @override
   void replace(DevToolsUrl other) {
-    ArgumentError.checkNotNull(other, 'other');
     _$v = other as _$DevToolsUrl;
   }
 
@@ -547,19 +462,12 @@ class DevToolsUrlBuilder implements Builder<DevToolsUrl, DevToolsUrlBuilder> {
   DevToolsUrl build() => _build();
 
   _$DevToolsUrl _build() {
-    final _$result =
-        _$v ??
-        new _$DevToolsUrl._(
+    final _$result = _$v ??
+        _$DevToolsUrl._(
           tabId: BuiltValueNullFieldError.checkNotNull(
-            tabId,
-            r'DevToolsUrl',
-            'tabId',
-          ),
-          url: BuiltValueNullFieldError.checkNotNull(
-            url,
-            r'DevToolsUrl',
-            'url',
-          ),
+              tabId, r'DevToolsUrl', 'tabId'),
+          url:
+              BuiltValueNullFieldError.checkNotNull(url, r'DevToolsUrl', 'url'),
         );
     replace(_$result);
     return _$result;
@@ -574,30 +482,20 @@ class _$DebugStateChange extends DebugStateChange {
   @override
   final String? reason;
 
-  factory _$DebugStateChange([
-    void Function(DebugStateChangeBuilder)? updates,
-  ]) => (new DebugStateChangeBuilder()..update(updates))._build();
+  factory _$DebugStateChange(
+          [void Function(DebugStateChangeBuilder)? updates]) =>
+      (DebugStateChangeBuilder()..update(updates))._build();
 
-  _$DebugStateChange._({
-    required this.tabId,
-    required this.newState,
-    this.reason,
-  }) : super._() {
-    BuiltValueNullFieldError.checkNotNull(tabId, r'DebugStateChange', 'tabId');
-    BuiltValueNullFieldError.checkNotNull(
-      newState,
-      r'DebugStateChange',
-      'newState',
-    );
-  }
-
+  _$DebugStateChange._(
+      {required this.tabId, required this.newState, this.reason})
+      : super._();
   @override
   DebugStateChange rebuild(void Function(DebugStateChangeBuilder) updates) =>
       (toBuilder()..update(updates)).build();
 
   @override
   DebugStateChangeBuilder toBuilder() =>
-      new DebugStateChangeBuilder()..replace(this);
+      DebugStateChangeBuilder()..replace(this);
 
   @override
   bool operator ==(Object other) {
@@ -659,7 +557,6 @@ class DebugStateChangeBuilder
 
   @override
   void replace(DebugStateChange other) {
-    ArgumentError.checkNotNull(other, 'other');
     _$v = other as _$DebugStateChange;
   }
 
@@ -672,19 +569,12 @@ class DebugStateChangeBuilder
   DebugStateChange build() => _build();
 
   _$DebugStateChange _build() {
-    final _$result =
-        _$v ??
-        new _$DebugStateChange._(
+    final _$result = _$v ??
+        _$DebugStateChange._(
           tabId: BuiltValueNullFieldError.checkNotNull(
-            tabId,
-            r'DebugStateChange',
-            'tabId',
-          ),
+              tabId, r'DebugStateChange', 'tabId'),
           newState: BuiltValueNullFieldError.checkNotNull(
-            newState,
-            r'DebugStateChange',
-            'newState',
-          ),
+              newState, r'DebugStateChange', 'newState'),
           reason: reason,
         );
     replace(_$result);