+++ dwds/test/dds_port_test.dart
@@ -8,6 +8,7 @@ library;
 
 import 'dart:io';
 
+import 'package:dwds/dwds.dart';
 import 'package:test/test.dart';
 import 'package:test_common/test_sdk_configuration.dart';
 
@@ -16,27 +17,44 @@ import 'fixtures/project.dart';
 import 'fixtures/utilities.dart';
 
 void main() {
-  late final TestSdkConfigurationProvider provider;
+  late TestSdkConfigurationProvider provider;
+  late TestContext context;
 
   setUp(() {
     provider = TestSdkConfigurationProvider();
+    context = TestContext(TestProject.test, provider);
   });
 
-  tearDown(() {
+  tearDown(() async {
+    await context.tearDown();
     provider.dispose();
   });
 
-  test('DWDS starts DDS with a specified port', () async {
-    final context = TestContext(TestProject.test, provider);
-
+  test('DWDS starts DDS with a specified port (deprecated)', () async {
     // Find a unused port for the test.
     final server = await HttpServer.bind(InternetAddress.loopbackIPv4, 0);
     final expectedPort = server.port;
     await server.close();
 
     await context.setUp(
-      debugSettings: TestDebugSettings.noDevTools().copyWith(
+      debugSettings: TestDebugSettings.noDevToolsLaunch().copyWith(
         ddsPort: expectedPort,
+      ),
+    );
+    expect(Uri.parse(context.debugConnection.ddsUri!).port, expectedPort);
+  });
+
+  test('DWDS starts DDS with a specified port', () async {
+    // Find a unused port for the test.
+    final server = await HttpServer.bind(InternetAddress.loopbackIPv4, 0);
+    final expectedPort = server.port;
+    await server.close();
+
+    await context.setUp(
+      debugSettings: TestDebugSettings.noDevToolsLaunch().copyWith(
+        ddsConfiguration: DartDevelopmentServiceConfiguration(
+          port: expectedPort,
+        ),
       ),
     );
     expect(Uri.parse(context.debugConnection.ddsUri!).port, expectedPort);