+++ dwds/test/debug_extension_test.dart
@@ -62,7 +62,7 @@ void main() async {
       group('Without encoding', () {
         setUp(() async {
           await context.setUp(
-            debugSettings: TestDebugSettings.withDevTools(
+            debugSettings: TestDebugSettings.withDevToolsLaunch(
               context,
             ).copyWith(enableDebugExtension: true, useSse: useSse),
           );
@@ -125,7 +125,7 @@ void main() async {
       group('With a sharded Dart app', () {
         setUp(() async {
           await context.setUp(
-            debugSettings: TestDebugSettings.withDevTools(
+            debugSettings: TestDebugSettings.withDevToolsLaunch(
               context,
             ).copyWith(enableDebugExtension: true, useSse: useSse),
           );
@@ -161,7 +161,7 @@ void main() async {
       group('With an internal Dart app', () {
         setUp(() async {
           await context.setUp(
-            debugSettings: TestDebugSettings.withDevTools(
+            debugSettings: TestDebugSettings.withDevToolsLaunch(
               context,
             ).copyWith(enableDebugExtension: true, useSse: false),
           );
@@ -231,7 +231,7 @@ void main() async {
   group('With encoding', () {
     setUp(() async {
       await context.setUp(
-        debugSettings: TestDebugSettings.noDevTools().copyWith(
+        debugSettings: TestDebugSettings.noDevToolsLaunch().copyWith(
           enableDebugExtension: true,
           urlEncoder:
               (url) async =>
@@ -263,7 +263,7 @@ void main() async {
     setUp(() async {
       await context.setUp(
         appMetadata: TestAppMetadata.externalApp().copyWith(hostname: 'any'),
-        debugSettings: TestDebugSettings.noDevTools().copyWith(
+        debugSettings: TestDebugSettings.noDevToolsLaunch().copyWith(
           enableDebugExtension: true,
         ),
       );